Got it pushing entries to the catalog; final version
Strategy won't work: embedding a consul agent (or many) in my process
would be hard or impossible (the agent isn't designed to be a lib).
Without serf participation, the service could be stale if I go offline,
etc.

The alternative is spinning up agents for each service, which is not
much more work than the old sidecar approach, or an Initializer